decomposition size is computed correctly for prime decompositions

+++ mpp/MPP-casestudy/src/helpers.c
@@ -279,8 +279,8 @@ struct pair get_decomposition_size(size_t dim_x, size_t dim_y, int comm_size)
 {
     struct pair sizes = get_decomposition_length(dim_x, dim_y, comm_size);
     struct pair dims;
-    dims.first = ceil(dim_x / sizes.first);
-    dims.second = ceil(dim_y / sizes.second);
+    dims.first = ceil(dim_x / (double) sizes.first);
+    dims.second = ceil(dim_y / (double) sizes.second);
 
     return dims;
 }
